+++ src/meta/highed.meta.options.extended.js
@@ -590,7 +590,18 @@ highed.meta.optionsExtended = {
               'areaspline'
             ],
             subTypeDefaults: {}
+          },
+          {
+            id: 'series-label--enabled',
+            pid: 'series.label.enabled',
+            defaults: 'true',
+            dataType: 'boolean'
           }
+          // {
+            // id: 'series-label--style',
+            // pid: 'series.label.style',
+            // dataType: 'font'
+          // }
         ]
       }
     ],